Fixed wanring
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@1149 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/Unittests/unittests_read_write_OM.cc
@@ -579,7 +579,7 @@ TEST_F(OpenMeshReadWriteOM, WriteTriangleFaceFloatProperty) {
 
     Mesh::FaceHandle f1 = Mesh::FaceHandle(0);
 
-    double va1ue1 = 3.1f;
+    float va1ue1 = 3.1f;
 
     mesh.property(prop,f1) = va1ue1;
